api/requirements.txt
anyio==3.7.1
autopep8==2.0.4
bcrypt==4.1.2
click==8.1.7
fastapi==0.109.2
h11==0.14.0
idna==3.6
Levenshtein==0.26.1
mypy==1.8.0
mypy-extensions==1.0.0
noise==1.2.2
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==1.10.13
PyJWT==2.8.0
python-multipart==0.0.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
textdistance==4.6.3
tomli==2.0.1
typing-extensions==4.9.0
uvicorn==0.27.1
websockets==12.0
